locations name not showing on admin panel
